[202405] Fix pre-commit install hook pathspec error (#21306)
What is the motivation for this PR?
202405 PR testing always failed at installing pre-commit hooks. The error is like:

[INFO] Initializing environment for https://github.com/sonic-net/sonic-mgmt.
An unexpected error has occurred: CalledProcessError: command: ('/usr/bin/git', 'checkout', '1.0.0+pre_commit')                 
return code: 1                                                                                                                  
stdout: (none)                                                                                                                  
stderr:                               
    error: pathspec '1.0.0+pre_commit' did not match any file(s) known to git
It's because the rev '1.0.0+pre_commit' tag for 'check-conditional-mark-sort' was removed in the sonic-mgmt repository.

How did you do it?
Update the rev to "202405" for this branch.