Issue 21185: [sflow] PTF script assumes port index map starts from 0 (#21186)
In the sflow PTF script, the port map is assumed to start at zero, which
is not always the case. The code calls 'self.dataplane.get_mac(0,0)' in
order to get a source port MAC. A few lines later, the code gets the
actual port to send the traffic, so we can use that source port MAC
instead (which is more correct anyway).

Signed-off-by: Yael Tzur <ytzur@nvidia.com>