Fix remove_dataacl_table fixture to correctly delete and restore DATAACL table (#21021)
What is the motivation for this PR?
As DATAACL was never restored for dualtor topologies, sometimes DATAACL table was bound to the same PortChannels that carry the BGP traffic, that lead to routes corruption and various tests failing.

How did you do it?
Modified the fixture to correctly check and delete and later accordingly restore the table

How did you verify/test it?
Ran test/test_acl_stress and verified the DATAACL was being restored in teardown

Signed-off-by: Dan Caugherty <dcaugher@cisco.com>