Skip swss events in telemetry/test_events.py for m* topos (#19872)
What is the motivation for this PR?
We were seeing failures while trying to verify the pfc storm event on m1 topos. This is because there is no QoS config. After inquiring with Microsoft we decided to skip this part of the test since QoS is not needed for the deployment.

How did you do it?
Refactored existing skip logic in the test which was being used for mx and m0 topos to include m1 and m2.

How did you verify/test it?
Ran test with and without change, test was passing with the change.

Signed-off-by: Yael Tzur <ytzur@nvidia.com>