Issue #21208: Enable sFlow for VS tests (#21209)
* Issue #21208: Enable sFlow for VS tests

Currently, sflow/test_sflow.py is disabled for virtual devices and is
not being tested. However, the test runs and passes just fine on virtual
devices when the sflow feature is enabled. Remove the skip condition and
enable sflow during the test when run on virtual devices.

* Remove hard-coded vs topology clause

This clause is not needed, since sflow should be enabled by default on
vs topos.

Signed-off-by: yenlu <yenlu@cisco.com>