Issue 22187: [sflow] Remove unnecessary 'config reload' call (#21190)
There is an extra 'config reload' call in the test_sflow.py test setup
code. It was added as a workaround for a bug that was filed in 2019, and
the workaround was supposed to have been removed in 2021, but the
commend and the 'config reload' call are still there. The test passes
fine without it, and it adds several minutes of overhead.

Signed-off-by: Priyansh Tratiya <ptratiya@microsoft.com>