[202411] Compatibility fixes for docker-sonic-mgmt based on Ubuntu 24.04 (#21045) (#21271) (#21286)
* Compatibility fixes for docker-sonic-mgmt based on Ubuntu 24.04 (#21045)

What is the motivation for this PR?
This PR is to continue the effort made by @yutongzhang-microsoft in #18339

The current docker-sonic-mgmt image is based on Ubuntu 20.04 which is end of support now. PR sonic-net/sonic-buildimage#24306 upgraded the base image of docker-sonic-mgmt to Ubuntu 24.04. Together, version of most packages are upgraded too. The upgrade introduced lots of compatibility issues. This PR is to fix all the compatibility issues.

All the fixes are backward compatible. The code change woks with both current and new docker-sonic-mgmt.

How did you do it?
Fix the snmp code caused by pysnmp upgrade.
Fix json dump of ansible result caused by pytest-ansible upgrade.
How did you verify/test it?
Take advantage of the current sonic-mgmt PR testing.
Verified that the code change works with new docker-sonic-mgmt in #20851
Verified that the code change works with the current docker-sonic-mgmt in this PR.

* Filter out unnecessary dataplane logs

PR https://github.com/sonic-net/sonic-mgmt/pull/21045 for fixing compatibility issues of docker-sonic-mgmt upgrade removed the hook for setting log levels of couple of loggers.

The code for setting log level of ansible loggers is no longer needed after previous ansible upgrade a few years ago.

However, setting log level for the 'dataplane' logger is still necessary.

This change added back the code to set log level for the 'dataplane' logger to avoid unnessary logs when 'ptfadapter' is used in test script.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>