Temporarily disable yang validation to unblock PR testing (#21389)
The test_pfcwd_all_port_storm.py has 2 issues:

1. It doesn't remove PFCWD configuration during tear down.
2. The detection_time is set to 200. Global POLL_INTERVAL is set to 400. According to YANG validation, detection_time should be greater than or equal to POLL_INTERVAL.The setting can't pass YANG validation.

Before the docker-sonic-mgmt image is upgraded, the run-yang-validation fixture is executed after  the core_dump_and_config_check.  The core_dump_and_config_check fixture performed a config reload because of issue 1. So, the yang validation executed after that is not able to detect the issue.

After the docker-sonic-mgmt image is upgraded, yang validation is executed before core_dump_and_config_check and surfaced the issue.

To unblock PR testing, this change temporarily disabled the autouse of the yang validation fixture.

After the issues of the pfcwd test are fixed, we need to revert this change.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: Priyansh Tratiya <ptratiya@microsoft.com>