This test exercises the console of a SONiC Device by pushing a large
block of data into the console as well as generate a large block of
of output out of the console to ensure that console remains stable.

This test supports all form factors including fixed and modular.
On a modular system (Currently only supports Arista, but can
easily be extend to to other systems), it exercises both the
supervisor and the linecard console.

This requires extending the related fixtures to create
a LinecardConsoleConn helper that can encapsulate the console
interaction on a linecard, which requires architecture specific
adaptation.

Signed-off-by: Patrick MacArthur <pmacarthur@arista.com>