override config tests: don't stop at first bad table
* tests/override_config_table/test_override_config_table.py
(load_minigraph_with_golden_empty_input): This function
asserts on the first discrepancy found. We instead gather the
names of the nonmatching tables into a list, and assert if the
list is not empty, showing the list in the assertion message.

* tests/override_config_table/test_override_config_table_masic.py:
Similar changes to multi-asic version of test
(load_minigraph_with_golden_empty_input): The difference is
that rather than gathering names of nonmatching tables,
we have (asic, tasble) tuples.