Update two test cases for Trixie environment (#21394)
* [T2] Switch to running reboot asynchronously and polling for device up

With Trixie, when running the reboot command, systemd might begin
killing user sessions before the reboot command fully returns to the
shell. This might mean that the SSH connection gets disconnected before
the command actually returns. Ansible will consider this as host
unreachable and treat it as a failure.

Work around this by running the reboot command asynchronously, ignoring
the result, and polling for the device to be up afterwards.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: Nishanth Sampath Kumar <nissampa@cisco.com>