Print stdout and stderr properly on a SSH command failure
Currently, when printing out the stdout and stderr on a SSH command
failure (run as part of tacacs tests), all that gets printed is that
it's a paramiko channel object with some metadata about it, which is not
helpful. Get the actual text and print it out.

The tricky part of this is that the `readlines()` method must be called
only when failing, because if `readlines()` is called on Paramiko, then
subsequent calls to the channel object will return different results,
since that data has been consumed. Because of that, wrap it in an
if-block so that it goes into that code path only when failing for
certain.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>