Auto assign reviewers GH action (#21475)
The implementation of the GitHub Action to automatically add reviewers to the PR based on the changed files and the code contributor statistics.

What is the motivation for this PR?
Improve the code change and review visibility

How did you do it?
Used the https://github.com/marketplace/actions/auto-assign-reviewer-by-files action.
Generated the yaml file with reviewers using the script
https://github.com/sonic-net/sonic-pipelines/tree/main/scripts/code-owners
and the helper scenario .code-reviewers/run.sh.
The reviewers are selected from the top contributors in the folder for the past 2 years
Implemented a github action triggered on the PR event.

How did you verify/test it?
Tested the approach in the private repo

Signed-off-by: nmirin <nikolay.a.mirin@gmail.com>
Signed-off-by: Yael Tzur <ytzur@nvidia.com>