Add "-n" to traceroute to speed up sanity check (#21443)
In sanity check, command `traceroute` is executed in the `print_logs` function to check PTF device reachability. However, the `traceroute` command is executed without the `-n` argument. Then it tries to resolve the DNS name of involved IP addresses. Usually the DNS resolve will timeout because PTF IP is usually not resolvable. Because of this, the `print_logs` function need more than 30 seconds to complete.

This change added "-n" argument to the traceroute command. With this change, usually 40 seonds less is required to run sanity check.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: selldinesh <dinesh.sellappan@keysight.com>