Update for dhcp and lag tests (#21499)
What is the motivation for this PR?
Update the loganalyzer ignore list to fix a regex issue, and update the GCU test to ignore new error syslogs.

The regex for ignoring systemd-networkd.socket failing to start included [1], which has a special meaning when interpreted as regex (match one character that has 1 in it; in other words, match a 1). Escape the brackets so that they get treated literally.

sonic-net/sonic-utilities#4105 started logging errors when a GCU command fails with a non-zero exit status. However, there is one case of a (known) failure where dhcp_relay fails to restart because it hit the start limit. These error logs need to be ignored.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>