Issue 20728: Enhance configlet/test_add_rack.py for ipv6-only topos
Currently, the test checks for both ipv4 and ipv6 neighbors
unconditionally, and will fail if either is missing. Change that
behavior to instead check to see if neighbors exist before running the
bgp checks. Add a failure case if no neighbors are found in either AF.
This way the test is AF-agnostic.

Signed-off-by: Anders Linn <anders@nexthop.ai>