Normalize vpp topo name (#21510)
What is the motivation for this PR?
vpp testbed has topo name with vpp suffix for the reason that in azure pipeline, it can use a different profile with different resources. However, from sonic-mgmt test perspective, t1-lag-vpp is exactly the same as t1-lag with the same topology. In sonic-mgmt tests, there are places check topo name. We don't want to add special logic for vpp topos. So we want to normalize the special topo name.

How did you do it?
In TestbedInfo, remove -vpp suffix when the testbed is loaded.

How did you verify/test it?
Run vpp sonic-mgmt test.

Any platform specific information?
Specific to vpp platform

Signed-off-by: Yue Gao <yuega2@cisco.com>
Signed-off-by: Nishanth Sampath Kumar <nissampa@cisco.com>