Increase timeout window for checking BGP Session State after restarting SWSS Container
On platforms with a large number of interfaces
running on larger topologies, the time taken for
SWSS and related services to come up can cause
the swss_docker subset of the bgp_session tests to
fail on timeout.

Increase the timeout window for BGP session checks
for the swss docker subtests.

Signed-off-by: Christopher Croy <ccroy@arista.com>