Fix COPP ARP Test
The COPP test for the ARP packet type does not
account for the fact that there are two ARP
trap IDs being acted upon in the target COPP trap
group: arp request and arp response.

Set the dst_ip to a non-reachable IP to avoid
generating an ARP response which would, depending
on the underlying trap rate calculation being
used, erroneously yield double the nominal rate.

Signed-off-by: Christopher Croy <ccroy@arista.com>