Update blocking-mode reboot test for Trixie (#21605)
What is the motivation for this PR?
Similar to #21394, the behavior of sudo reboot is changed with Trixie, in that the SSH connection may get disconnected before the sudo reboot command completes with an exit code. This will mean that Ansible will consider the host to be unreachable and throw an exception.

How did you do it?
The common reboot infra that we have runs sudo reboot in its own thread. That means that the HostUnreachable exception that Ansible raises will get raised in that thread, but won't affect the main thread; this, in a way, handles this particular behavior change. Therefore, for the blocking reboot test case, reuse that common infra when running on a Trixie image and we need to reboot the device (this probably can be done unconditionally, but make it conditional for now).

How did you verify/test it?
Tested with a subset of T0 KVM tests: https://elastictest.org/scheduler/testplan/6936fccd392767e9bf67ecb8

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: Yael Tzur <ytzur@nvidia.com>