Fix test_power_off_reboot to properly power on subsets of PSUs
This fix is meant to restore the test functionality where power on
sequences are subsets of all PSUs instead restoring power on all PSUs
at once every time.

It also fixes the duplicate entries in power_on_seq_list when there is
only one PSU. That is, when there is only 1 PSU we shouldn't do the
subset of PSUs followed by all PSUs since that is the same thing.

Signed-off-by: Peter <peterbailey@arista.com>