[action] [PR:21582] fix: only disable route check for T2 (#21613)
Change the temporarily_disable_route_check fixture logic to only apply to T2 topology for now.

What is the motivation for this PR?
The current disable-and-enable routeCheck monitor logic is causing test flakiness on some non-T2 platforms (see #16876 (comment)). Certain platforms require additional time to restart the routeCheck monitor, which can leave it inactive when the next test begins and result in false failures. We would like to address this issue urgently in this PR.

In a follow-up PR, I will properly enhance the temporarily_disable_route_check fixture so that:

Users can choose which topologies apply the disable-and-enable routeCheck behavior
The fixture uses a wait_until() timeout to verify the routeCheck status is as expected before proceeding to the next step
How did you do it?
How did you verify/test it?
I ran the updated login on a non-T2 platform (Mx) and can confirm it's working well:
https://elastictest.org/scheduler/testplan/693272f7392767e9bf67e930
image

I also verified the logic on T2 platform and can confirm it's still having this logic: https://elastictest.org/scheduler/testplan/6932767fbcc3fac23371a83c
image

Signed-off-by: Chenyang Wang <chenyangw233@gmail.com>
Co-authored-by: Chenyang Wang <49756587+cyw233@users.noreply.github.com>