Allow non-zero exit code when enabling a systemd service
With Trixie, service files for SONiC containers are runtime-generated
files (because of how systemd-sonic-generator now has to work). This
means that enabling such a service will result in an error being
returned from systemd saying that the service is a generated service and
cannot be explicitly enabled.

Allow such errors for this command; functionally, things still work.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>