Remove redundant error checking for commands
If a command fails, and `module_ignore_errors` isn't set to `True`, then
an exception would be raised, and the test would fail. The additional
error check afterwards isn't doing anything, and in the case of a
command having `module_ignore_errors` set to `True`, doesn't ignore the
failure.

Remove that block of code and simplify the logic a bit.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>