Allow non-zero exit code when enabling a systemd service (#21693)
* Allow non-zero exit code when enabling a systemd service

With Trixie, service files for SONiC containers are runtime-generated
files (because of how systemd-sonic-generator now has to work). This
means that enabling such a service will result in an error being
returned from systemd saying that the service is a generated service and
cannot be explicitly enabled.

Allow such errors for this command; functionally, things still work.

If a command fails, and `module_ignore_errors` isn't set to `True`, then
an exception would be raised, and the test would fail. The additional
error check afterwards isn't doing anything, and in the case of a
command having `module_ignore_errors` set to `True`, doesn't ignore the
failure.

Remove that block of code and simplify the logic a bit.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: Yael Tzur <ytzur@nvidia.com>