test_fib.py: test_hash test case fails when IP protocol is set to GRE (47) (#18750)
test_fib.py: test_hash test case fails when IP protocol is set to GRE (47).
Test constructs a packet with IP protocol GRE but the entire packet is not formed correctly to be a valid
GRE packet. For that reason it gets dropped at the kernel on the receiving server where test verifies packet
reception.
Test already skips some of the protocols: IPinIP, Ipv6InIpv4 protocols which get dropped if malformed.
Need to add GRE as well to the list.

What is the motivation for this PR?
test_fib.py: test_hash test case fails without the fix

How did you do it?
Added GRE to the list of skipped IP protocols.

How did you verify/test it?
Test passed with the fix.

Signed-off-by: selldinesh <dinesh.sellappan@keysight.com>