[Thinkit]: Modifications for arbitration test
The set of changes required for the arbitration test to work. This goes with https://github.com/sonic-net/sonic-pins/pull/1670 commited in sonic-pins.

In pins_deps.bzl, the sonic-pins is upgraded to the latest. This required upgrading of a few packages. The arbitration test required the addition of the util files and the upgrade of googletest. The gutil directory has moved in the sonic-pins depo. The new path is updated in several files.

Signed-off-by: Sree Iyer <sbiyer@google.com>
Signed-off-by: Sreemoolanathan Iyer <sbiyer@google.com>