Fix test_dscp_to_queue_mapping_uniform_mode for t1-isolated-d128 (#21663)
This test selects a downlink as a source port and an uplink as a dest port, but the t1-isolated-d128 has no uplinks. This change merges the list of down/uplink ports and selects one at random as the source port, and cleans up some incorrect assertions and unused arguments.

This is a manual backport of #21662

Signed-off-by: Rick Robbins <rick@arista.com>