Add subscribe sample testcase verifying received timestamps
Summary:
This adds a simple subscribe sample testcase. This test exercises the
fix for a bug where on every notification received by the telemetry
binary the timer was reset. Before the fix was merged, this testcase
would fail with context deadline exceeded as we stopped receiving
updates. With the product fix the test case now passes.

It changes the helper funcs to allow overridable origin and target
attributes.

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [x] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

---------

Signed-off-by: Ronan Mac Fhlannchadha <ronan@nexthop.ai>