Merged PR 17773: Updated Management IP for Arista Chassis Supervisor in SVC Lab
The supervisor previously used the IP address 10.206.144.21; however, we encountered conflicts indicating that this IP was already in use by another device. We were unable to identify the conflicting device, as it does not appear in either the sonic-mgmt-int repository or the Networking Metadata repository. As a result, we updated the supervisor IP address to 10.206.144.49.