Modify ntpd conf to prevent socket exhaustion
When running ntpd on the ptf_ardut container as a part of ntp testing,
modify the config to ignore all interfaces except mgmt.

It was found while developing multi-vrf that it is possible at scale to
run out of sockets for ntp, which prevents the service from starting
correctly.  Listening on the mgmt interface only allows the test to run
ntp from the ptf_ardut container to the dut correctly, and ntp tests
pass.

Signed-off-by: Will Rideout <wrideout@arista.com>