[manual] [PR:21614] fix: enhance routeCheck toggle logic (#21673)
Further enhance the routeCheck monitor disable-and-enable logic:

Users can choose which topologies apply the disable-and-enable routeCheck behavior
Use wait_until() timeout to verify the routeCheck status is as expected before proceeding to the next step

What is the motivation for this PR?
This is a follow-up PR of #21582. Not all platforms need the "temporarily disable roureCheck monitor" feature, and the routeCheck monitor will take some time to startup after running sudo monit start routeCheck on some platforms. Therefore, we want to allow the users to choose which topologies they want to apply the disable-and-enable routeCheck behavior (Now only T2, LT2 & UT2 are allowed). Besides, we added a wait_until() timeout to verify the routeCheck status is as expected before proceeding to the next step.

How did you do it?
How did you verify/test it?
I verified it on a T0 platform and I can confirm this logic will be skipped:

Signed-off-by: Chenyang Wang <chenyangw233@gmail.com>